requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless
segment-anything
pyyaml
matplotlib
streamlit
Pillow
numpy
